change delta time to be long not float

+++ gamecode/src/com/saints/gamecode/Arena.java
@@ -51,7 +51,7 @@ public class Arena {
 
     }
     //Gets called from the game loop when the arena should update
-    public void update(float delta){
+    public void update(long delta){
         characterController.update(delta);
     }
 }
